+++ app/(root)/(routes)/(homepage)/filmPosterDetail/filmPosterDetail.css
@@ -1,3 +1,15 @@
+.genreTag {
+  background: rgba(255, 255, 255, 0.2);
+  backdrop-filter: blur(30px);
+  border-radius: 40px;
+  padding: 2px 8px 2px 8px;
+  box-shadow: 0 4px 6px rgba(0, 0, 0, 0.1);
+  border: 2px solid rgba(255, 255, 255, 0.2);
+  margin-right: 16px;
+  margin-bottom: 16px;
+  display: inline-block;
+}
+
 .filmButton {
     width: 206px;
     height: 54px;
@@ -45,8 +57,8 @@
   
 
 .NFTClaimButton {
-    width: 114px;
-    height: 53px;
+    width: 6rem;
+    height: 3rem;
     border-radius:20px;
     border: 1px solid rgba(255, 255, 255, 0.2);
     box-shadow: 0px 4px 14px 0px rgba(0, 0, 0, 0.16);
