+++ app/(root)/(routes)/(homepage)/index.tsx
@@ -10,6 +10,32 @@ import {LiveFilmSection} from './LiveFilmSection';
 
 import {LiveFilms, films,  FilmData, FilmCardInRowProps, sectionFilmRow, listOfSectionFilmRows, SectionFilmRowsPopularOfWeek} from './data';
 
+import { FilmPosterDetail } from './filmPosterDetail';
+
+const filmPosterDetailData = {
+  posterSrc: '/assets/images/film1.png',
+  title: 'Wednesday',
+  logoSrc: '/assets/images/Logo-Wednesday.png',
+  description: "Wednesday Addams, a teenager who possesses psychic powers.Wednesday's cold, emotionless personality and her defiant nature make it difficult for her to connect with her schoolmates and cause her to run afoul of the school's principal Larissa Weems. However, she discovers she has inherited her mother's psychic abilities which allow her to solve a local murder mystery.",
+  rating: 8.2,
+  duration: 130,
+  releaseDate: 'Sep 23 2023',
+  genres: ['Drama', 'Adventure'], 
+
+  stars: 'Lewis Tan, Jessica McNamee, Josh Lawson',
+  director: 'Simon McQuoid',
+
+  NFTClaimImg: './assets/NFTs/NFT4.png',
+  NFTEventName: 'Wednesday',
+
+  expirationDate: '2023-10-18T12:30:00',
+
+  trailerVideo: 'link vid from youtube',
+  trailerImg: '/assets/images/film1.png',
+
+  eventImg: './assets/filmDetail/gallery/galleryImg1.png',
+}
+
 export const HomePage = ()=> {
     return (
       <div className=""
@@ -18,7 +44,7 @@ export const HomePage = ()=> {
           // border: '2px solid black',
         }}
       >
-        <FilmDetails/>
+        {/* <FilmDetails/>
         {
           SectionFilmRowsPopularOfWeek.map((sectionFilmRow: sectionFilmRow) => (
             <FilmRow
@@ -27,7 +53,29 @@ export const HomePage = ()=> {
               filmRow={sectionFilmRow.filmRow}
             />
           ))
-        }
+        } */}
+
+        <FilmPosterDetail
+          posterSrc={filmPosterDetailData.posterSrc}
+          title={filmPosterDetailData.title}
+          logoSrc={filmPosterDetailData.logoSrc}
+         
+         
+          duration={filmPosterDetailData.duration} 
+          releaseDate={filmPosterDetailData.releaseDate}
+          genres={filmPosterDetailData.genres}
+          stars={filmPosterDetailData.stars}
+          director={filmPosterDetailData.director}
+
+          NFTClaimImg={filmPosterDetailData.NFTClaimImg}
+          NFTEventName={filmPosterDetailData.NFTEventName}
+          expirationDate={filmPosterDetailData.expirationDate}
+
+          trailerVideo={filmPosterDetailData.trailerVideo}
+          trailerImg={filmPosterDetailData.trailerImg}
+          eventImg={filmPosterDetailData.eventImg}
+
+        />
 
         <LiveFilmSection
